refactor: move shared UI resources (Type, PaddingValues) from login to components
